rename crate to cargo-credential-libsecret
