--- credential/cargo-credential-gnome-secret/src/main.rs
@@ -1,5 +1,0 @@
-//! Cargo registry gnome libsecret credential process.
-
-fn main() {
-    cargo_credential::main(GnomeSecret);
-}
