+++ credential/cargo-credential-libsecret/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "cargo-credential-gnome-secret"
+name = "cargo-credential-libsecret"
 version = "0.3.1"
 edition.workspace = true
 license.workspace = true
@@ -10,4 +10,3 @@ description = "A Cargo credential process that stores tokens with GNOME libsecre
 anyhow.workspace = true
 cargo-credential.workspace = true
 libloading.workspace = true
-
