+++ src/cargo/util/auth/mod.rs
@@ -451,7 +451,7 @@ fn credential_action(
             "cargo:token-from-stdout" => Box::new(BasicProcessCredential {}),
             "cargo:wincred" => Box::new(cargo_credential_wincred::WindowsCredential {}),
             "cargo:macos-keychain" => Box::new(cargo_credential_macos_keychain::MacKeychain {}),
-            "cargo:libsecret" => Box::new(cargo_credential_gnome_secret::GnomeSecret {}),
+            "cargo:libsecret" => Box::new(cargo_credential_libsecret::LibSecretCredential {}),
             process => Box::new(CredentialProcessCredential::new(process)),
         };
         config.shell().verbose(|c| {
